dehaze-algorithm/requirements.txt
addict
future
lmdb~=1.4.1
numpy~=1.26.4
opencv-python~=4.9.0.80
Pillow~=10.2.0
pyyaml~=6.0.1
requests~=2.31.0
scikit-image~=0.22.0
scipy~=1.11.4
torch>=1.7
torchvision~=0.16.0+cu121
tqdm~=4.65.0
yapf
pyiqa==0.1.4
einops~=0.7.0
timm~=0.6.13

h5py~=3.10.0
natsort~=8.4.0
natten~=0.15.1+torch210cu121
thop~=0.1.1-2209072238
torchsummary~=1.5.1
matplotlib~=3.8.3
setuptools~=59.5.0
